Refactored folder structure, group import bug fix
The following things were done in this commit

* The OverlayEditor and the complicated mess in `club_view.dart` was done
  away with given the design change.
* Rebase with current refactored folder directory. For some reason all
  my code disappeared in thin air and a views folder got created in
  src/app. Had to manually copy paste all the code into the ui/views.
* There is a bug in injectable which does not work well with group
  imports. This was fixed by making the auth service a lazy singleton.
