--- elaichi/lib/src/app/widgets/club/club_appbar.dart
@@ -1,13 +1,0 @@
-import 'package:elaichi/ui/views/club/club_viewmodel.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
-import 'package:flutter/src/widgets/framework.dart';
-import 'package:stacked/stacked.dart';
-
-class ClubAppBar extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    // TODO: implement build
-    ;
-  }
-}
