+++ elaichi/lib/src/ui/views/viewmodels.dart
@@ -1,3 +1,4 @@
+export 'club/club_viewmodel.dart';
 export 'feed/feed_viewmodel.dart';
 export 'home/home_viewmodel.dart';
 export 'startup/startup_viewmodel.dart';
