+++ elaichi/lib/src/ui/views/views.dart
@@ -1,3 +1,4 @@
+export 'club/club_view.dart';
 export 'feed/feed_view.dart';
 export 'home/home_view.dart';
 export 'pageview/elaichi_page_view.dart';
