+++ elaichi/lib/src/ui/widgets/club/stories_archive/club_story_grid_item.dart
@@ -1,5 +1,5 @@
 import 'package:easy_localization/easy_localization.dart';
-import 'package:elaichi/datamodels/api_models.dart';
+import 'package:elaichi/datamodels.dart';
 import 'package:flutter/material.dart';
 
 /// Individual item shown in story archive grid in club view.
