+++ elaichi/test/viewmodel_tests/club_viewmodel_test.dart
@@ -1,11 +1,10 @@
 import 'dart:math';
 
 import 'package:dartz/dartz.dart';
-import 'package:elaichi/app/failure.dart';
-import 'package:elaichi/app/locator.dart';
-import 'package:elaichi/datamodels/datamodels.dart';
-import 'package:elaichi/services/api.dart';
-import 'package:elaichi/ui/views/club/club_viewmodel.dart';
+import 'package:elaichi/core.dart';
+import 'package:elaichi/datamodels.dart';
+import 'package:elaichi/services.dart';
+import 'package:elaichi/viewmodels.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
